Moved nearby_nodes function to client/src/graphics/draw.rs
